refactor(NxDPreTrainedModel): regroup NeuronPreTrainedModel methods
